New Gateway Update 1
-Wk Savage- This update adds a new gateway level, the abandonedship.dmm.
The map is a ship that was used for research by nanotrasen to study the
far reaches of space. What you dont know is that even though all seems
well, there could be danger around every corner. This update also adds
new unsimed walls, thanks paracode, as well as adding the base for an
emergency light that requires no power and will be used on the new
gateway and possibly future gateway maps.